Make pages for Random and Picker components
